refactor(ui): extract AdvanceMixin for rank-up checks + dialog launchers
Move the rank advancement nicebar checks and the dialog launchers
they trigger out of L5RMain into l5r/ui/advance.py:

* act_choose_skills
* check_rank_advancement, check_school_new_spells, check_free_kihos,
  check_new_skills, check_affinity_wc
* learn_next_school_spells, learn_next_free_kiho
* show_advance_rank_dlg, show_buy_skill_dlg, show_buy_emph_dlg
* show_select_affinity, show_select_deficiency

These methods are tightly coupled (each check_* shows a nicebar whose
button triggers the matching show_/learn_ launcher) so they ship as a
single mixin.

main.py: 2170 -> 1979 lines.
